requirements.txt
fastapi
uvicorn
typing
google-cloud-logging
selenium
webdriver_manager
pandas